backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database (SQLite - no external dependencies)
sqlalchemy==2.0.23
alembic==1.12.1
# psycopg2-binary==2.9.9  # Optional - only needed for PostgreSQL

# Document Processing
PyPDF2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
Pillow==10.1.0

# OCR (Optional - comment out if you don't need OCR)
# pytesseract==0.3.10
# opencv-python==4.8.1.78

# NLP and AI (Open Source)
spacy==3.7.2
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2
datasets==2.14.6
# Optional API clients (commented out)
# openai==1.3.5
# anthropic==0.7.7

# Background Tasks
celery==5.3.4
redis==5.0.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
aiofiles==23.2.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Monitoring and Logging
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0
